Bot +1

+++ app/Utils/Telegram/Commands/CheckinCommand.php
@@ -61,7 +61,7 @@ class CheckinCommand extends Command
             // 回送信息
             $response = $this->replyWithMessage(
                 [
-                    'text'       => '您未绑定本站账号，您可以进入网站的 **资料编辑**，在右下方绑定您的账号.',
+                    'text'       => Config::get('user_not_bind_reply'),
                     'parse_mode' => 'Markdown',
                 ]
             );
@@ -78,12 +78,21 @@ class CheckinCommand extends Command
                 ]
             );
         }
-        // 消息删除任务
-        TelegramTools::DeleteMessage([
-            'chatid'      => $ChatID,
-            'messageid'   => $response->getMessageId(),
-            'executetime' => (time() + Config::get('delete_message_time'))
-        ]);
+        if ($ChatID < 0) {
+            // 消息删除任务
+            TelegramTools::DeleteMessage([
+                'chatid'      => $ChatID,
+                'messageid'   => $response->getMessageId(),
+                'executetime' => (time() + Config::get('delete_message_time'))
+            ]);
+            if (Config::get('enable_delete_user_cmd') === true) {
+                TelegramTools::DeleteMessage([
+                    'chatid'      => $ChatID,
+                    'messageid'   => $Message->getMessageId(),
+                    'executetime' => (time() + Config::get('delete_message_time'))
+                ]);
+            }
+        }
         return;
     }
 }
